紐付けのバグ修正

+++ app/Http/Controllers/LinkController.php
@@ -36,8 +36,12 @@ class LinkController extends Controller
     // 新規作成実行(記事：質問＝１：多)
     public function postDocumentToQuestions(Request $request,Document $document)
     {
-        $document->questions()->detach($request['detach_id']);
-        $document->questions()->attach($request['attach_id']);
+        if($request['detach_id']){
+            $document->questions()->detach($request['detach_id']);
+        }
+        if($request['attach_id']){
+            $document->questions()->attach($request['attach_id']);
+        }
         return redirect('/links/index');
     }
     
@@ -63,8 +67,13 @@ class LinkController extends Controller
     // 新規作成実行(記事：質問＝多：1)
     public function postQuestionToDocuments(Request $request,Question $question)
     {
-        $question->documents()->detach($request['detach_id']);
-        $question->documents()->attach($request['attach_id']);
+        // dd($request);
+        if($request['detach_id']){
+            $question->documents()->detach($request['detach_id']);
+        }
+        if($request['attach_id']){
+            $question->documents()->attach($request['attach_id']);
+        }
         return redirect('/links/index');
     }
 }
